Astro 6 update; Cloudflare Fixes; Add speaking link
