+++ package.json
@@ -24,10 +24,10 @@
   },
   "homepage": "https://github.com/domkirby/domkirby-com#readme",
   "dependencies": {
-    "@astrojs/cloudflare": "^12.6.12",
+    "@astrojs/cloudflare": "^13.1.3",
     "@astrojs/tailwind": "^6.0.2",
     "@tailwindcss/typography": "^0.5.19",
-    "astro": "^5.17.3",
+    "astro": "^6.0.8",
     "tailwindcss": "^3.4.19"
   },
   "devDependencies": {
