+++ src/components/Header.astro
@@ -4,6 +4,7 @@ const currentPath = Astro.url.pathname;
 const navLinks = [
   { label: 'Home', href: '/' },
   { label: 'Blog', href: '/blog' },
+  { label: 'Speaking', href: 'https://speaking.domkirby.com', external: true },
   { label: 'Store', href: 'https://store.domkirby.com', external: true },
   { label: 'About Me', href: '/about' },
   { label: 'Tools and Stuff', href: '/tools' },
