--- src/content/config.ts
@@ -1,15 +1,0 @@
-import { defineCollection, z } from 'astro:content';
-
-const posts = defineCollection({
-  type: 'content',
-  schema: z.object({
-    title: z.string(),
-    date: z.coerce.date(),
-    categories: z.array(z.string()).default([]),
-    tags: z.array(z.string()).default([]),
-    coverImage: z.string().optional(),
-    draft: z.boolean().default(false),
-  }),
-});
-
-export const collections = { posts };
