+++ wrangler.jsonc
@@ -1,5 +1,5 @@
 {
-  "main": "dist/_worker.js/index.js",
+  "main": "@astrojs/cloudflare/entrypoints/server",
   "name": "domkirby-com",
   "compatibility_date": "2025-09-27",
   "compatibility_flags": [
